requirements.txt
# Core Dependencies
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.10.6
python-dateutil==2.8.2

# Development Dependencies
pytest==8.3.4
pytest-cov==6.0.0
pytest-mock==3.14.0
black==24.10.0
flake8==7.1.1
mypy==1.14.1
